* Fix markdown issue with renderHardBreak's * Fix issue when tapping on a push notification * Fix splash screen dismissal
