+++ app/constants/index.js
@@ -4,11 +4,9 @@
 import NavigationTypes from './navigation';
 import StorageTypes from './storage';
 import ViewTypes from './view';
-import SplashScreenTypes from './splash_screen';
 
 export {
     NavigationTypes,
-    SplashScreenTypes,
     StorageTypes,
     ViewTypes
 };
