--- app/constants/splash_screen.js
@@ -1,8 +1,0 @@
-// Copyright (c) 2017 Mattermost, Inc. All Rights Reserved.
-// See License.txt for license information.
-
-import keyMirror from 'mattermost-redux/utils/key_mirror';
-
-export default keyMirror({
-    CLOSE: null
-});
